Small refactors, renaming and files re-ordering
