+++ public/index.html
@@ -6,7 +6,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400;700&display=swap" rel="stylesheet">
-    <title>Pharma Inc</title>
+    <title>Pharma Inc.</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
