+++ src/App.jsx
@@ -1,9 +1,9 @@
+import GlobalStyles from "./styles/global";
 import Routes from "./routes";
 import Header from "./components/Header";
 import { ListUsersProvider } from "./contexts/listUsersContext";
 import { FiltersProvider } from "./contexts/filtersContext";
 
-import GlobalStyles from "./styles/global";
 
 export function App() {
   return (
