+++ src/components/UserList/UserModal/index.tsx
@@ -5,8 +5,9 @@ import {
   DialogContent,
   DialogActions,
   Button,
+  Box,
 } from "@mui/material";
-import { useListUsersContext } from "../../contexts/listUsersContext";
+import { useListUsersContext } from "../../../contexts/listUsersContext";
 import { useHistory, useParams } from "react-router-dom";
 
 export default function UserModal() {
@@ -14,6 +15,10 @@ export default function UserModal() {
   const { id: paramID } = useParams<{ id: string }>();
   const { usersData } = useListUsersContext();
   const selectedUser = usersData.find(({ login: { id }}) => id === paramID);
+
+  if (!selectedUser) return null;
+
+  const { name, gender, email, dob, nat, picture } = selectedUser;
   
   const onClose = () => {
     history.push('/');
@@ -23,22 +28,20 @@ export default function UserModal() {
     <Dialog open={!!selectedUser} onClose={onClose}>
       <DialogTitle>Scientist Details</DialogTitle>
       <DialogContent>
-        {selectedUser && (
-          <div>
+          <Box>
             <img
-              src={selectedUser.picture.large}
+              src={picture.large}
               alt="Scientist"
               style={{ maxWidth: "100%", marginTop: "1rem" }}
             />
             <Typography>
-              Name: {`${selectedUser.name.first} ${selectedUser.name.last}`}
+              Name: {`${name.first} ${name.last}`}
             </Typography>
-            <Typography>Gender: {selectedUser.gender}</Typography>
-            <Typography>Email: {selectedUser.email}</Typography>
-            <Typography>Date of Birth: {selectedUser.dob.date}</Typography>
-            <Typography>Nationality: {selectedUser.nat}</Typography>
-          </div>
-        )}
+            <Typography>Gender: {gender}</Typography>
+            <Typography>Email: {email}</Typography>
+            <Typography>Date of Birth: {dob.date}</Typography>
+            <Typography>Nationality: {nat}</Typography>
+          </Box>
       </DialogContent>
       <DialogActions>
         <Button onClick={onClose}>Close</Button>
