+++ src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {CssBaseline} from '@mui/material';
+
 import App from './App';
 
 ReactDOM.render(
