+++ src/routes/index.tsx
@@ -1,7 +1,7 @@
-import { Switch, Route, BrowserRouter } from "react-router-dom";
+import { Route, BrowserRouter } from "react-router-dom";
 
 import UserList from "../components/UserList";
-import UserModal from "../components/UserModal";
+import UserModal from "../components/UserList/UserModal";
 
 export default function Routes() {
   return (
